requirements.txt
-e .
boto3
aws_cdk.aws_apigateway
aws_cdk.aws_lambda
aws_cdk.aws_dynamodb